+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog references the relevant changes (bug and security fixes) for the 
     * Fixed bug in authorization policy the swapped inherence and knowledge factors.
     * Fixed bug where response headers were not being properly validated and verified.
     * Added verification of webhook headers.
+    * Added dynamic authorization request title
 
   * 4.2.1